+++ Runtime/NodeTypes/ActionNode.cs
@@ -1,0 +1,4 @@
+namespace BTDesigner
+{
+    public abstract class ActionNode : Node {}
+}