+++ Runtime/NodeTypes/CompositionNode.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace BTDesigner
+{
+    public abstract class CompositionNode : Node
+    {
+        public List<Node> children = new();
+    }
+}