+++ Runtime/NodeTypes/UtilityNode.cs
@@ -1,0 +1,7 @@
+namespace BTDesigner
+{
+    public abstract class UtilityNode : Node
+    {
+        public Node child;
+    }
+}