+++ Runtime/UtilityNodes/RepeaterNode.cs
@@ -1,0 +1,15 @@
+namespace BTDesigner
+{
+    public class RepeaterNode : UtilityNode
+    {
+        protected override void OnStart() { }
+
+        protected override NodeState OnUpdate()
+        {
+            child.Update();
+            return NodeState.Running;
+        }
+
+        protected override void OnEnd() { }
+    }
+}